+++ ProgramZaRacunovodstvo/Models/Prodaja.xaml.cs
@@ -11,10 +11,10 @@ namespace ProgramZaRacunovodstvo.Models
         public string BrojFakture { get; set; } = string.Empty;
         public string Status { get; set; } = string.Empty;
         public string Primalac { get; set; } = string.Empty;
-        public double Osnovica { get; set; }
+        public decimal Osnovica { get; set; }
 
-        public double Pdv { get; set; }
-        public double Ukupno { get; set; }
+        public decimal Pdv { get; set; }
+        public decimal Ukupno { get; set; }
         public DateOnly DatumSlanja { get; set; }
 
         public string OsnovicaRSD => $"{Osnovica:N2} RSD";
